Changed note editor to QPlainTextEdit and improved Esc event handler.
When Esc is pressed in the CaptureDialog, the warning msg box will not
show if no title and notes are entered (this is nicer).

It makes more sense to use a plain text edit. Why have formatting if
we're not going to print anything? It will also simplify evolution as
its easier to migrate from plain text to rich text/html than the other
way around.