Made editing possible in RecallView.
Original intent was to keep this view and its model read only. However,
it's far quicker to edit through the available widgets directly than to
call up an edit dialog each time.

Seeing as capture/recall speed is the project's primary goal, the
complexity introduced in this change is deemed acceptable. All hail.