Fixed index mapping and implemented button state switching.
Previous revision did not map QModelIndexes between from the SFProxy to
the ItemModel. This is now corrected.

Several refactorings improve the encapsulation of ItemModel, such that
it is responsible for connecting signals and slots that have to do with
serializing for itself (previously, the view did this which was messy).
Furthermore, it signals when an item changes and when items are saved,
thus allowing easy state maintenance for enabling/disabling the save
button in RecallView.

NOTE: This revision temporarily removes the tag combo box in RecallView.
The improved regexp filtering via the SFProxy deprecates the old filter
implementation. Several new ideas will be prototyped.