Basic encryption and decryption features implemented.
The features rely on the GPGME_CryptoWrapper in recap.h (which in turn
relies on GPG being installed and GPG agent being configured properly).
That said, items can be encrypted in capture mode and summarily
decrypted in recall mode now.

All dialogs required to select an encryption key and/or warn of errors
are implemented. Also, further settings are added to the config file to
remember such choices across sessions.